extraction/requirements.txt
pdfplumber==0.7.6
pymongo==4.6.0
redis==4.5.5
python-dateutil==2.8.2
tqdm==4.65.0
